request formatter util
